+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added support for `mr-manager -v` to print the installed version and exit.
+
 ## [0.0.2] - 2026-04-01
 
 ### Added